+++ doc/version_history.rst
@@ -17,6 +17,7 @@ v1.20.0
 * Add new ``maintel/m1m3/raise_m1m3.py`` to raise MainTel M1M3 mirror. 
 * Add new ``laser_tracker/set_up.py`` script to set up and turn on the laser tracker.
 * Add new ``laser_tracker/shut_down.py`` script to switch off the laser tracker.
+* Add new ``laser_tracker/align.py`` script to align mtcs with laser tracker.
 
 v1.19.2
 -------